Add option to disable bind-mounting /exe
In some of the jobs we'd like to sandbox with sio2jail,
eg. compilation jobs, the executable we want to run is a part of
the sandbox, instead of being provided separately.

In such cases, the default behaviour of treating 0th argument as an
outside path and bind-mounting it to /exe is unnecessary and actually
makes things more difficult.

Add an option to disable this implicit bind mount
(and any implicit bind-mounts that may be added in the future).