Don't confuse exit status with kill signal in OutputBuilder
It used to be that:
(a) setting nonzero kill signal would set exit status to singal + 128
(b) setting exit status > 128 would set kill signal to status - 128

However, we want to be able to differentiate between a program
getting killed and a program that fakes getting killed
by calling exit(128+signal)

Therefore, stop doing (b). Leave (a) in place.

Also, remove both of the above calcuations from any places other than
OutputBuilder.

Now if killSignal is nonzero, we can be sure the program did recive
that signal and didn't handle it,
and if it's zero, we can be sure the program did exit(exitStatus).

We still keep case (a) so that someone only looking at exitStatus
doesn't get fooled into tinking the program exitted succesfully
when in fact it died from an unhandled signal.