Fix incompatibility with OITT output format
In the first numeric field of the result line, we used to print
actual exit code of the program. However, that's not what OITT does.

The number OITT returns encodes the RE/TLE/MLE/etc. status, as well as
kill signal or program's actual exit status. It does so in a way that
wouldn't fit in 8 bits, but it doesn't matter, since it's an int printed
in a text form, so full range of int can be used.

Change OITimeToolOutputBuilder to print an encoded status code
the same wayt OITT does it.