Fix mode checking in `open` syscall filter
The syscall filter for open(2) checks that the mode & O_RDWR == 0.
Looks like someone assumed that O_RDWR == O_WRONLY | O_RDONLY,
which is not the case.
In fact O_WRONLY & O_RDWR == 0, which means the filter allows opening
files WRONLY. This is wrong.

Fix it by changing the check to mode & (O_RDWR | O_WRONLY) == 0