Build libcap when it isn't present
Download and build (actually just downloading would be sufficient since
all we need are headers and we aren't linking with libcap library) the
libcap library when it isn't found on host (or when user asks to do so).
A simple patch (partially copied from gentoo) is needed to build it on
newer systems. Since we aren't linking sio2jail with libcap we use
ADD_CUSTOM_TARGET and ADD_DEPENDENCIES to build libcap before sio2jail,
later it could be changed to ADD_LIBRARY(libcap STATIC IMPORTED), just
like it's done with libseccomp.