Add support for heterogenous cpus to PerfListener
This is achieved by opening a raw perf event for every cpu core type,
quite similiarly to the perf userspace tool.
The values needed to open the proper raw events are gathered from
/sys/devices/cpu*/{type,events/instructions,format/*}
This solution should work as long as the sysfs interface doesn't
drastically change while maintaining compatibility with older kernels,
as the mentioned interface seems to be the same as in at least 2014.